pyproject.toml
[tool.poetry]
authors = ["gladson <gladsonbrito@gmail.com>"]
description = ""
name = "sig"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
beautifulsoup4 = "^4.12.2"
python = "^3.10"
python-decouple = "^3.8"
requests = "^2.31.0"
selenium = "^4.12.0"
undetected-chromedriver = "^3.5.3"
html5lib = "^1.1"
webdriver-manager = {url = "https://github.com/jnhyperion/webdriver_manager/archive/refs/heads/add-file-lock.zip"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
types-requests = "^2.31.0.2"
types-selenium = "^3.141.9"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
